Fix warnings in util
util/charset/unidata.h(304): warning: integer conversion resulted in a change of sign
util/generic/hash_primes.h:30:54: warning: static_assert with no message is a C++1z extension [-Wc++1z-extensions]

ref:274d09854e7b4130cf931407b35fc221fffb0aba

+++ util/charset/unidata.h
@@ -301,7 +301,7 @@ inline wchar32 ToTitle(wchar32 ch) {
 
 inline int ToDigit(wchar32 ch) {
     ui32 i = NUnicode::NPrivate::CharInfo(ch);
-    return (i & IS_DIGIT) ? (i >> SVAL_OFFSET) : -1;
+    return (i & IS_DIGIT) ? static_cast<int>(i >> SVAL_OFFSET) : -1;
 }
 
 // BIDI properties
